add src folder to project structure
