+++ src/index.ts
@@ -1,0 +1,2 @@
+export { default as Calendar } from "./calendar";
+export type { IYear, IMonth, IDay, Days, Months } from "./interfaces";
